refactor(cli): extract conversation region from App
