+++ packages/cli/src/chat/tui/panes/ConversationRegion.tsx
@@ -1,0 +1,275 @@
+// Conversation render boundary: static scrollback, live transcript, foreground
+// surfaces, compact side panels, tips, and queued follow-ups above the footer.
+
+// Third-party imports
+import { Box } from 'ink';
+import { useLayoutEffect, useRef, useState, type ReactNode } from 'react';
+
+// Local imports - shared constants and schemas
+import { isActiveStatus } from '@common/constants/streamStatus';
+import { type ActiveChildInfo, type StreamTabId } from '@shared/schemas';
+import { clamp } from '@utils/core';
+
+// Local imports - conversation panes and layout
+import {
+  allocateMiddleRows,
+  allocateSidePanelRows,
+  PINNED_CHROME_ROWS,
+  shouldShowTipRow,
+  shouldShowTodosPlanPanel,
+  staticScrollbackTarget,
+  staticTranscriptRowBudget,
+} from '../appLayout';
+import {
+  visibleSubagentRows,
+  type ChildStreamEntries,
+} from '../state/childExecutions';
+import {
+  isScopedTranscriptViewport,
+  transcriptViewportChange,
+  transcriptViewportKey,
+  type TranscriptViewportChange,
+} from '../state/transcriptViewportMode';
+import { clampModalWidth } from '../ui/theme';
+import { ConversationPane } from './ConversationPane';
+import {
+  QueuedFollowUpsPanel,
+  queuedFollowUpPanelRowCount,
+} from './QueuedFollowUpsPanel';
+import { StaticConversationTranscript } from './StaticConversationTranscript';
+import { SubagentList, subagentPanelRowCount } from './SubagentList';
+import { TipRow } from './TipRow';
+import { TodosPlanPanel, todosPlanPanelRowCount } from './TodosPlanPanel';
+import type { ForegroundSurfaceKind } from '../appInteractionPolicy';
+import type { ChildControlStreamTarget } from '../state/childControls';
+import type { StreamSlice } from '../state/cliState';
+
+// Cap the bottom subagent/todos panels so they never crowd out the
+// conversation or push the input bar off-screen.
+const BOTTOM_PANEL_MAX_ROWS = 10;
+const EMPTY_SUBAGENT_ROWS: readonly ActiveChildInfo[] = [];
+
+interface ConversationRegionSnapshot {
+  readonly activeStreamId: StreamTabId | undefined;
+  readonly childStreamEntries: ChildStreamEntries;
+  readonly foregroundMaxRows: number | undefined;
+  readonly foregroundKind: ForegroundSurfaceKind | undefined;
+  readonly parentStream: ReadonlyMap<StreamTabId, StreamTabId>;
+  readonly reverseSearchOpen: boolean;
+  readonly rootStreamId: StreamTabId | undefined;
+  readonly slashPaletteOpen: boolean;
+  readonly streamTabsVisible: boolean;
+  readonly streams: ReadonlyMap<StreamTabId, StreamSlice>;
+  readonly subagentPanelTarget: ChildControlStreamTarget;
+  readonly transcriptViewerStreamId: StreamTabId | undefined;
+}
+
+interface ConversationRegionProps {
+  readonly agentSelectionAvailable: boolean;
+  readonly children: (queuedFollowUpPanelVisible: boolean) => ReactNode;
+  readonly colorEnabled?: boolean;
+  readonly columns: number;
+  readonly onTranscriptViewportChange?: (
+    change: TranscriptViewportChange,
+  ) => void;
+  readonly renderForegroundSurface: (
+    availableRows: number,
+    transcriptWidth: number,
+  ) => ReactNode;
+  readonly rows: number;
+  readonly snapshot: ConversationRegionSnapshot;
+}
+
+export function ConversationRegion({
+  agentSelectionAvailable,
+  children,
+  colorEnabled,
+  columns,
+  onTranscriptViewportChange,
+  renderForegroundSurface,
+  rows,
+  snapshot,
+}: ConversationRegionProps): React.JSX.Element {
+  const [tipHour] = useState(() => new Date().getHours());
+  const transcriptViewerOpen = snapshot.transcriptViewerStreamId !== undefined;
+  const foregroundOpen = snapshot.foregroundKind !== undefined;
+  const inputBarVisible = !foregroundOpen;
+  const viewportKey = transcriptViewportKey({
+    activeStreamId: snapshot.activeStreamId,
+    parentStream: snapshot.parentStream,
+    transcriptViewerStreamId: snapshot.transcriptViewerStreamId,
+  });
+  const scopedTranscript = isScopedTranscriptViewport(viewportKey);
+  const scrollbackTarget = staticScrollbackTarget({
+    activeStreamId: snapshot.activeStreamId,
+    rootStreamId: snapshot.rootStreamId,
+    scopedTranscript,
+  });
+  const previousViewportKey = useRef<string | undefined>(undefined);
+
+  useLayoutEffect(() => {
+    const previous = previousViewportKey.current;
+    previousViewportKey.current = viewportKey;
+    const change = transcriptViewportChange({
+      previousViewportKey: previous,
+      nextViewportKey: viewportKey,
+    });
+    if (change) onTranscriptViewportChange?.(change);
+  }, [onTranscriptViewportChange, viewportKey]);
+
+  const activeSlice = snapshot.activeStreamId
+    ? snapshot.streams.get(snapshot.activeStreamId)
+    : undefined;
+  const activeResponseRunning = isActiveStatus(activeSlice?.status);
+  const queuedFollowUpMessages = activeSlice?.queuedFollowUpMessages ?? [];
+  const queuedFollowUpPanelWanted =
+    !foregroundOpen && queuedFollowUpMessages.length > 0;
+  const footerRows =
+    PINNED_CHROME_ROWS.status +
+    (inputBarVisible ? PINNED_CHROME_ROWS.input : 0) +
+    (snapshot.streamTabsVisible ? PINNED_CHROME_ROWS.streamTabsWorstCase : 0);
+  const requestedQueuedFollowUpPanelRows = queuedFollowUpPanelWanted
+    ? queuedFollowUpPanelRowCount(queuedFollowUpMessages)
+    : 0;
+  const queuedFollowUpPanelRows = queuedFollowUpPanelWanted
+    ? clamp(rows - footerRows, 0, requestedQueuedFollowUpPanelRows)
+    : 0;
+  const queuedFollowUpPanelVisible = queuedFollowUpPanelRows > 0;
+  const tipRowVisible =
+    !scopedTranscript &&
+    shouldShowTipRow({
+      foregroundOpen,
+      hasQueuedFollowUps: queuedFollowUpPanelWanted,
+    });
+  const staticTranscriptRows = scopedTranscript
+    ? undefined
+    : staticTranscriptRowBudget({
+        footerRows,
+        foregroundOpen,
+        queuedFollowUpPanelRows,
+        rows,
+        tipVisible: tipRowVisible,
+      });
+  const subagentPanelTarget = snapshot.subagentPanelTarget;
+  const subagentPanelRows = subagentPanelTarget.streamId
+    ? visibleSubagentRows(
+        subagentPanelTarget.streamId,
+        snapshot.childStreamEntries,
+        snapshot.streams,
+      )
+    : EMPTY_SUBAGENT_ROWS;
+  const hasSubagentPanel = !foregroundOpen && subagentPanelTarget.hasItems;
+  const hasTodosPlanPanel = shouldShowTodosPlanPanel({
+    foregroundOpen,
+    hasPlan: activeSlice?.plan != null,
+    status: activeSlice?.status,
+    todos: activeSlice?.todos ?? [],
+  });
+  const transcriptWidth = clampModalWidth(columns);
+  const { foregroundRows, transcriptRows } = allocateMiddleRows({
+    foregroundMaxRows: snapshot.foregroundMaxRows,
+    foregroundOpen,
+    inputVisible: inputBarVisible,
+    queuedFollowUpPanelRows,
+    reverseSearchOpen: snapshot.reverseSearchOpen,
+    reserveTranscriptRows: snapshot.foregroundKind !== 'transcript',
+    rows,
+    slashPaletteOpen: snapshot.slashPaletteOpen,
+    streamTabsVisible: snapshot.streamTabsVisible,
+    staticTranscriptRows: staticTranscriptRows ?? 0,
+    tipVisible: tipRowVisible,
+  });
+  // The subagent/todos panels live at the bottom of the same vertical column.
+  // Reserve only as many rows as the panels actually need, capped so they
+  // never take more than half the transcript or push the input off-screen.
+  const subagentContentRows =
+    hasSubagentPanel && subagentPanelTarget.slice
+      ? subagentPanelRowCount(
+          subagentPanelRows,
+          subagentPanelTarget.slice.activeProcesses,
+        )
+      : 0;
+  const todosPlanContentRows =
+    hasTodosPlanPanel && activeSlice
+      ? todosPlanPanelRowCount(activeSlice.todos, activeSlice.plan)
+      : 0;
+  const bottomPanelBudget = Math.min(
+    BOTTOM_PANEL_MAX_ROWS,
+    subagentContentRows + todosPlanContentRows,
+    Math.floor(transcriptRows / 2),
+  );
+  const conversationRows = transcriptRows - bottomPanelBudget;
+  const { subagentRows, todosPlanRows } = allocateSidePanelRows({
+    subagentContentRows,
+    todosPlanContentRows,
+    rows: bottomPanelBudget,
+  });
+  const foregroundSurface = renderForegroundSurface(
+    foregroundRows,
+    transcriptWidth,
+  );
+
+  return (
+    <>
+      {transcriptViewerOpen ? null : (
+        <StaticConversationTranscript
+          colorEnabled={colorEnabled}
+          maxRows={staticTranscriptRows}
+          ownerKey={scrollbackTarget.ownerKey}
+          scrollbackStreamId={scrollbackTarget.streamId}
+          width={transcriptWidth}
+        />
+      )}
+      <Box flexDirection="column">
+        <Box flexDirection="column" overflowY="hidden">
+          {!transcriptViewerOpen && conversationRows > 0 ? (
+            <ConversationPane
+              colorEnabled={colorEnabled}
+              width={transcriptWidth}
+              maxRows={conversationRows}
+            />
+          ) : null}
+          {foregroundSurface ? (
+            // Cap the modal area at its row budget but size to the surface's
+            // actual content. Each foreground surface already windows itself
+            // to the budget, so a fixed height leaves dead rows below it.
+            <Box
+              flexDirection="column"
+              maxHeight={foregroundRows}
+              alignItems="flex-start"
+              overflowY="hidden"
+            >
+              {foregroundSurface}
+            </Box>
+          ) : null}
+        </Box>
+        {bottomPanelBudget > 0 ? (
+          <Box flexDirection="column" overflowY="hidden">
+            <SubagentList
+              maxRows={subagentRows}
+              subagents={subagentPanelRows}
+              activeProcesses={subagentPanelTarget.slice?.activeProcesses}
+              processOutput={subagentPanelTarget.slice?.processOutput}
+            />
+            <TodosPlanPanel maxRows={todosPlanRows} />
+          </Box>
+        ) : null}
+        {tipRowVisible ? (
+          <TipRow
+            agentSelectionAvailable={agentSelectionAvailable}
+            hour={tipHour}
+            responseRunning={activeResponseRunning}
+          />
+        ) : null}
+        {queuedFollowUpPanelVisible ? (
+          <QueuedFollowUpsPanel
+            maxRows={queuedFollowUpPanelRows}
+            messages={queuedFollowUpMessages}
+            width={columns}
+          />
+        ) : null}
+        {children(queuedFollowUpPanelVisible)}
+      </Box>
+    </>
+  );
+}
